fixed unlinkdirkey

+++ rootfs/usr/local/bin/unlinkdirkey.sh
@@ -6,5 +6,5 @@ CDIR=$1
 /usr/sbin/e4crypt get_policy "${CDIR}" | grep "Error getting policy" > /dev/null 2>&1 && echo "policy error" && exit
 POLICY=$(/usr/sbin/e4crypt get_policy "${CDIR}" | cut -d':' -f2)
 KEYID=$(keyctl show | grep ${POLICY} | cut -d'-' -f1)
-keyctl unlink ${KEYID}
+[ ! -z "${KEYID}" ] && keyctl unlink ${KEYID}
 echo 3 > /proc/sys/vm/drop_caches
